+++ M3/Utils/SProt/sprot_l.h
@@ -70,7 +70,7 @@ void sprot_init_fifo(sprot_fifo* fifo);
 /**
  * Must be called in loop to execute commands.
  */
-void process_fifo(sprot_fifo* fifo, sprot_efunc table[], void (*default_fun)(sprot_buff_entry*), uint8_t tbl_entries);
+void process_fifo(sprot_fifo* fifo, sprot_efunc table[], void (*default_fun)(sprot_buff_entry*), uint16_t tbl_entries);
 
 /**
  * Must be called in loop to timeout incoming commands.
